Ensure that all tasks exit
I've found that on the odd occasion that the cli takes an extended
period of time to exit.

After a fair amount of testing I can confirm that all of the
`async-done` handles do correctly complete, however, upon using
[wtfnode](https://www.npmjs.com/package/wtfnode) I found that there are
a few results that are keeping the cli open.

> [WTF Node?] open handles:
> - Sockets:
>   - undefined:undefined -> undefined:undefined
>     - Listeners:
>   - undefined:undefined -> undefined:undefined
>     - Listeners:
> - Timers:
>   - (30000 ~ 30 s) bound cleanupWebsocketResources @ $PWD/node_modules/engine.io/node_modules/ws/lib/Sender.js:59

Adding `exit(0)` resolves these issues.