format-error.js: Prefer `e.error.message` if available
Previously when using `e.error` and it was an object, we got just an "[object Object]" message in the resulting Error object (and error message in terminal). With this patch instead when there is a `message` property inside `e.error`, use that to get a better error report, and falling back to old behaviour in other cases.